Fixar så att man kan logga in i vårt API

+++ St-dafint.ab/src/main/java/com/example/nicecleaning/dto/CreateCleanDTO.java
@@ -6,7 +6,7 @@ public record CreateCleanDTO(
         String optionalMessage,
         int id) {
 
-    public int appUserId() {
+    public long appUserId() {
         return id;
     }
 }
